requirements.txt
scikit-learn
pandas
flask
joblib
requests
reportlab
pdfkit
opencv-python
pillow
numpy
